requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
aiosqlite>=0.19.0

# Authentication
PyJWT>=2.8.0

# HTTP client for NVR discovery
aiohttp>=3.9.0

# WebRTC support (optional - for sub-second latency)
aiortc>=1.6.0
av>=10.0.0

# Optional: for production
# gunicorn>=21.2.0
